+++ parm/wflow/verify_pre.yaml
@@ -187,7 +187,7 @@ metatask_ASCII2nc_obs:
 task_run_MET_Point2Grid_GOES:
   <<: *default_task_verify_pre
   attrs:
-    cycledefs: cycledefs_obs_days_inst
+    cycledefs: forecast
     maxtries: '2'
   command: "&JOBSdir;/POINT2GRID.sh"
   memory: '{{  verification_resources.execution.point2grid.memory if user.MACHINE not in ["NOAACLOUD"] }}'
@@ -203,22 +203,22 @@ task_run_MET_Point2Grid_GOES:
         value: '&GOESAOD_OBS_DIR;'
       datadep_adp:
         value: '&GOESADP_OBS_DIR;'
-      or:
-        not:
-          taskvalid:
-            attrs:
-              task: 'get_obs_goes_aod'
-        taskdep:
-          attrs:
-            task: 'get_obs_goes_aod'
-      or:
-        not:
-          taskvalid:
-            attrs:
-              task: 'get_obs_goes_adp'
-        taskdep:
-          attrs:
-            task: 'get_obs_goes_adp'
+      # Check that the flag files that indicate that the get_obs_ccpa tasks
+      # are complete are all present before launching any PcpCombine task.
+      sh_goesaod_obs_available:
+        attrs:
+          shell: /bin/bash
+        command: find {{ workflow.WFLOW_FLAG_FILES_DIR }} -name "get_obs_goesaod_*_complete.txt" -exec basename {} \; 2>/dev/null |
+          grep -oE '[0-9]{8}' |
+          awk -F/ -v s={{ workflow.OBS_DAYS_ALL_CYCLES_CUMUL[0] }} -v e={{ workflow.OBS_DAYS_ALL_CYCLES_CUMUL[-1] }} '($NF >= s && $NF <= e)' |
+          test $(wc -l) -ge {{ workflow.OBS_DAYS_ALL_CYCLES_CUMUL|length }}
+      sh_goesadp_obs_available:
+        attrs:
+          shell: /bin/bash
+        command: find {{ workflow.WFLOW_FLAG_FILES_DIR }} -name "get_obs_goesadp_*_complete.txt" -exec basename {} \; 2>/dev/null | 
+          grep -oE '[0-9]{8}' | 
+          awk -F/ -v s={{ workflow.OBS_DAYS_ALL_CYCLES_CUMUL[0] }} -v e={{ workflow.OBS_DAYS_ALL_CYCLES_CUMUL[-1] }} '($NF >= s && $NF <= e)' | 
+          test $(wc -l) -ge {{ workflow.OBS_DAYS_ALL_CYCLES_CUMUL|length }}
 
 metatask_PcpCombine_APCP_all_accums_obs_CCPA:
   var:
