+++ ush/render_metplus_confs.py
@@ -1,0 +1,69 @@
+#!/usr/bin/env python3
+import argparse
+import logging
+import os
+import sys
+from datetime import datetime, timedelta
+from jinja2 import Environment, FileSystemLoader
+
+def render_metplus_confs(cfg,settings,template_fn,vx_leadhr_list,tasks):
+    """Renders metplus conf files from the appropriate template and user settings.
+    If VX_TASKS > 1 and vx_leadhr_list > 1, renders a conf file for each parallel task.
+    Returns the filename(s) of metplus conf files that were rendered"""
+
+    logger = logging.getLogger(__name__)
+
+    num_fhrs = len(vx_leadhr_list)
+    outconfs = []
+    logger.debug(f"Loading METplus conf template file: {template_fn}")
+    logger.debug(f"from directory {cfg['user']['METPLUS_CONF']}")
+    env = Environment(loader=FileSystemLoader(cfg['user']['METPLUS_CONF']))
+    template = env.get_template(template_fn)
+
+    if tasks > 1:
+        # Break down forecast hours according to number of tasks requested
+        if tasks > num_fhrs:
+            logger.warning("Number of tasks is greater than number of forecast hours\n"\
+                           f"Only running {num_fhrs} tasks in parallel")
+            tasks = len(vx_leadhr_list)
+
+        for i in range(tasks):
+            logger.debug(f"Rendering conf file for task {i}")
+            # We will have i conf files, so append i to the base filename for each
+            settings['metplus_log_fn'] = f"{settings['metplus_log_fn'].rsplit('.',1)[0]}.{i}"
+            settings['metplus_config_fn'] = f"{settings['metplus_config_fn'].rsplit('.',1)[0]}.{i}"
+            outconf = f"{settings['output_dir']}/{settings['metplus_config_fn']}"
+            logger.debug(f"metplus log file for task: {settings['metplus_log_fn']}")
+            logger.debug(f"metplus final rendered conf for task: {outconf}")
+            hours_per_task,remainder = divmod(num_fhrs,tasks)
+            # For cases where things don't divide evenly, ensure we get best distribution
+            if i >= remainder:
+                vx_leadhr_list, task_fhrs = vx_leadhr_list[hours_per_task:],vx_leadhr_list[:hours_per_task]
+            else:
+                vx_leadhr_list, task_fhrs = vx_leadhr_list[hours_per_task+1:],vx_leadhr_list[:hours_per_task+1]
+            settings['vx_leadhr_list'] = ', '.join(map(str,task_fhrs))
+            logger.debug(f"Task {i} will process lead hours: {settings['vx_leadhr_list']}")
+            rendered = template.render(settings)
+            with open(outconf,'w', encoding="utf-8") as f:
+                f.write(rendered)
+            outconfs.append(outconf)
+
+    else:
+        #Remove task-specific suffixes if we're only using one task
+        settings['metplus_log_fn'] = settings['metplus_log_fn'].rsplit('.',1)[0]
+        settings['metplus_config_fn'] = settings['metplus_config_fn'].rsplit('.',1)[0]
+        outconf = f"{settings['output_dir']}/{settings['metplus_config_fn']}"
+        logger.debug("Rendering conf file")
+        logger.debug(f"metplus log file: {settings['metplus_log_fn']}")
+        logger.debug(f"metplus final rendered conf: {settings['metplus_config_fn']}")
+        logger.debug(f"Will process lead hours: {settings['vx_leadhr_list']}")
+        rendered = template.render(settings)
+        with open(outconf,'w', encoding="utf-8") as f:
+            f.write(rendered)
+        outconfs = [outconf]
+
+    return outconfs
+
+if __name__ == "__main__":
+
+    raise RuntimeError("This is not a python script, just a function definition")
